// packages/services/shopify/src/__generated__/graphql.ts
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  Color: { input: any; output: any; }
  DateTime: { input: any; output: any; }
  Decimal: { input: any; output: any; }
  HTML: { input: any; output: any; }
  ISO8601DateTime: { input: any; output: any; }
  JSON: { input: any; output: any; }
  URL: { input: any; output: any; }
  UnsignedInt64: { input: any; output: any; }
};
// ...
export type VideoSource = {
  readonly __typename: 'VideoSource';
  /** The format of the video source. */
  readonly format: Scalars['String']['output'];
  /** The height of the video. */
  readonly height: Scalars['Int']['output'];
  /** The video MIME type. */
  readonly mimeType: Scalars['String']['output'];
  /** The URL of the video. */
  readonly url: Scalars['String']['output'];
  /** The width of the video. */
  readonly width: Scalars['Int']['output'];
};